Update ILogger dependency in RequestLoggingBehavior
Updated the `ILogger` dependency in the `RequestLoggingBehavior` class. The generic type parameter `TRequest` was previously used as the type argument for `ILogger`. It now uses the `RequestLoggingBehavior<TRequest, TResponse>` class itself as the type argument for `ILogger`. The constructor of `RequestLoggingBehavior` class has also been updated to accept an `ILogger<RequestLoggingBehavior<TRequest, TResponse>>` instead of `ILogger<TRequest>`.