Encryption (#29)
* Fix formatting in ValidationBehavior.cs

* Added "Mediat" to the list of namespaces in settings.json

* Updated Azure provider and added new resources in main.tf

Updated the Azure provider configuration in the main.tf file to include key vault features. Added new resources `azurerm_app_configuration_key` and `azurerm_key_vault_key` for storing the encryption key URL and creating an RSA encryption key respectively.

* Updated package and added IEncryption interface

Updated the `Microsoft.FeatureManagement` package reference in the `Neba.Application.csproj` file. Added a new namespace `Neba.Application.Security` and a new interface `IEncryption` in the `IEncryption.cs` file. The `IEncryption` interface includes two asynchronous methods: `EncryptAsync` and `DecryptAsync`, both accepting a nullable string `data` and a `CancellationToken` `cancellationToken`.

* Created Encryption class and corresponding tests

Implemented a new `Encryption` class in `Encryption.cs` that uses Azure's KeyVault `CryptographyClient` to perform encryption and decryption operations. The class implements the `IEncryption` interface and includes `EncryptAsync` and `DecryptAsync` methods. The `EncryptAsync` method encrypts a string using the RSA-OAEP encryption algorithm and the `DecryptAsync` method decrypts a base64 string.

Also added a new `EncryptionTests` class in `EncryptionTests.cs` to test the functionality of the `Encryption` class. The tests validate the encryption and decryption process, ensuring that the original value is returned when a plaintext value is encrypted and then decrypted, different values are returned when the same value is encrypted multiple times, an empty string is returned when the input is null or empty, and the original text is returned when different encrypted values from the same original text are decrypted.

* Make sealed

* Subject: Added Encryption class and corresponding unit tests

Body:
This commit introduces a new `Encryption` class in the `Neba.UI.Services` namespace, implementing the `IEncryption` interface. The class uses RSA encryption with a 2048-bit key, generated from a `JsonWebKey` object. The `Encrypt` method returns an encrypted string, or an empty string for null or empty input.

Additionally, a new `EncryptionTests` class has been added to the `Neba.UI.UnitTests.Services` namespace, providing unit tests for the `Encryption` class. The tests validate the encryption process, ensuring different outputs for the same input, and correct handling of null or empty strings.

Minor formatting changes have also been made in `Neba.UI.UnitTests.csproj`, specifically adjusting the closing tag of the `FrameworkReference` element.

* Added Neba.FunctionalTests project and updated visibility

Added a new project "Neba.FunctionalTests" to the solution "Neba.sln" and created a corresponding project file "Neba.FunctionalTests.csproj". Updated "Neba.Infrastructure" and "Neba.UI" projects to make their internal types visible to the new project. Also, added a new test class "EncryptionTests" to verify encryption and decryption functionality between UI and API layers.

* Update package versions in Directory.Packages.props

* Update ILogger dependency in RequestLoggingBehavior

Updated the `ILogger` dependency in the `RequestLoggingBehavior` class. The generic type parameter `TRequest` was previously used as the type argument for `ILogger`. It now uses the `RequestLoggingBehavior<TRequest, TResponse>` class itself as the type argument for `ILogger`. The constructor of `RequestLoggingBehavior` class has also been updated to accept an `ILogger<RequestLoggingBehavior<TRequest, TResponse>>` instead of `ILogger<TRequest>`.

* Updated encryption setup and welcome message

Updated the `Home.razor` file to change the welcome message. Added new namespaces and `AddEncryption` method in `InfrastructureConfiguration.cs` to retrieve a key from Azure Key Vault and add an instance of `IEncryption` to the service collection. Updated `AddKeyVault` method in `Program.cs` to return a `KeyClient` instance. Updated `appsettings.Development.json` and `appsettings.json` files to include a new `Encryption` section. Added two new Azure App Configuration keys in `main.tf` for the encryption key name and URL, and updated the existing key for the Key Vault URL.

* Add encryption service support to application

The `appsettings.json` file was updated to include a new "Encryption" section with a "Url" field for specifying the URL of an encryption service. The `InfrastructureDependencyInjection.cs` file was updated to include cryptography and security namespaces, and a new `AddEncryption` method was added to the `InfrastructureDependencyInjection` class. This method retrieves the encryption URL from the configuration, creates a new `CryptographyClient`, and adds an `Encryption` instance to the service collection. This update suggests that the application now supports encryption functionality, with the encryption service being configured based on the application's configuration settings.

* Refactor `_mockLogger` type and instantiation

The type of `_mockLogger` has been refactored from `ILogger<TestRequest>` to `ILogger<RequestLoggingBehavior<TestRequest, ErrorOr<TestResponse>>>`, making it specific to the `RequestLoggingBehavior` class. The instantiation of `_mockLogger` in the `RequestLoggingBehaviorTests` constructor has been updated accordingly.

* Update Neba.Application.csproj file