Update SonarCloud workflow to install DotNet Coverage tool (#9)
This pull request includes an update to the GitHub Actions workflow to
install a new tool for code coverage. The most important change is the
addition of a step to install `dotnet-coverage` if it is not already
cached.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/step-sonarcloud.yml`](diffhunk://#diff-32f4348b589a90c1eb1bbcb13d93ba3d0711e774890bc36d1b94f1106ffcfbe5R53-R56):
Added a step to install `dotnet-coverage` if it is not already cached.